Do not tile Troll Canyon entrance.
The entrance (32x64) is less than a single world map tile (50x50).
Tiling does not have an advantage, and trolls only attack when a player
is nearby anyways. This makes the map less confusing for map editors and
DMs trying to reset the map.

Also take this opportunity to correct capitalization and make the bridge
troll actually say something (was broken due to typo).

git-svn-id: https://svn.code.sf.net/p/crossfire/code/maps@18837 282e977c-c81d-0410-88c4-b93c2d0d6712